fix: correção integer para long id parceiro relacionado a eventos

+++ src/main/java/br/org/larescolaredencao/dto/AtualizarEventoDTO.java
@@ -18,7 +18,7 @@ public class AtualizarEventoDTO {
     private BigDecimal valor;
     private TipoEvento tipoEvento;
     private String comentarioPosEvento;
-    private List<Integer> parceirosIds;
+    private List<Long> parceirosIds;
 
     public String getTitulo() {
         return titulo;
@@ -84,11 +84,11 @@ public class AtualizarEventoDTO {
         this.comentarioPosEvento = comentarioPosEvento;
     }
     
-    public List<Integer> getParceirosIds() {
+    public List<Long> getParceirosIds() {
         return parceirosIds;
     }
 
-    public void setParceirosIds(List<Integer> parceirosIds) {
+    public void setParceirosIds(List<Long> parceirosIds) {
         this.parceirosIds = parceirosIds;
     }
 }
